# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "seo-content-agent",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/strategy": {
      "get": {
        "tags": [
          "strategy"
        ],
        "summary": "Get Strategy",
        "operationId": "get_strategy_api_v1_projects__project_id__strategy_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StrategyResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "strategy"
        ],
        "summary": "Update Strategy",
        "operationId": "update_strategy_api_v1_projects__project_id__strategy_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StrategyResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/strategy/versions": {
      "get": {
        "tags": [
          "strategy"
        ],
        "summary": "List Strategy Versions",
        "operationId": "list_strategy_versions_api_v1_projects__project_id__strategy_versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StrategyVersionListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/strategy/versions/{version}": {
      "get": {
        "tags": [
          "strategy"
        ],
        "summary": "Get Strategy Version",
        "operationId": "get_strategy_version_api_v1_projects__project_id__strategy_versions__version__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_StrategyVersionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_StrategyResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategyResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[StrategyResponse]"
      },
      "ApiResponse_StrategyVersionListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategyVersionListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[StrategyVersionListResponse]"
      },
      "ApiResponse_StrategyVersionResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategyVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[StrategyVersionResponse]"
      },
...
      "AudienceSchema": {
        "properties": {
          "segments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Segments",
            "description": "Target segments"
          },
          "personas": {
            "items": {
              "$ref": "#/components/schemas/PersonaSchema"
            },
            "type": "array",
            "title": "Personas",
            "description": "Target personas"
          },
          "needs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Needs",
            "description": "Core customer needs"
          },
          "buying_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Buying Stages",
            "description": "Stages in buying journey"
          }
        },
        "type": "object",
        "title": "AudienceSchema"
      },
...
      "BusinessContextSchema": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name",
            "description": "Business / brand name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Overview of business",
            "default": ""
          },
          "industry": {
            "type": "string",
            "title": "Industry",
            "description": "Primary industry",
            "default": ""
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locations",
            "description": "Physical/service locations"
          }
        },
        "type": "object",
        "title": "BusinessContextSchema"
      },
...
      "CompetitorSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Competitor name",
            "default": ""
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Competitor domain",
            "default": ""
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths",
            "description": "Competitor strengths"
          }
        },
        "type": "object",
        "title": "CompetitorSchema"
      },
...
      "ErrorItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorItem",
        "description": "Stable machine-readable error entry."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Market or country name",
            "default": ""
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Country/region code",
            "default": ""
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "description": "Primary market",
            "default": false
          }
        },
        "type": "object",
        "title": "MarketSchema"
      },
...
      "PersonaSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Persona name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Persona summary",
            "default": ""
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems",
            "description": "Problems faced"
          },
          "goals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Goals",
            "description": "Goals and objectives"
          },
          "funnel_stage": {
            "type": "string",
            "title": "Funnel Stage",
            "description": "Target funnel stage",
            "default": "TOFU"
          }
        },
        "type": "object",
        "title": "PersonaSchema"
      },
      "ProductServiceSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Product or service name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Short description",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "product or service",
            "default": "service"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Landing page URL if known"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "Business priority (1 is highest)",
            "default": 1
          }
        },
        "type": "object",
        "title": "ProductServiceSchema"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "description": "Request metadata shared by success and failure responses."
      },
...
      "SEOGoalSchema": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Goal type",
            "default": "LEAD_GENERATION"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Goal details",
            "default": ""
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "Goal priority (1 is highest)",
            "default": 1
          }
        },
        "type": "object",
        "title": "SEOGoalSchema"
      },
...
      "StrategyDataSchema-Input": {
        "properties": {
          "business_context": {
            "$ref": "#/components/schemas/BusinessContextSchema"
          },
          "audience": {
            "$ref": "#/components/schemas/AudienceSchema"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductServiceSchema"
            },
            "type": "array",
            "title": "Products"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ProductServiceSchema"
            },
            "type": "array",
            "title": "Services"
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/MarketSchema"
            },
            "type": "array",
            "title": "Markets"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/SEOGoalSchema"
            },
            "type": "array",
            "title": "Goals"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/CompetitorSchema"
            },
            "type": "array",
            "title": "Competitors"
          },
          "seo_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seo Objectives"
          },
          "content_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Objectives"
          },
          "priority_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Topics"
          }
        },
        "type": "object",
        "title": "StrategyDataSchema"
      },
      "StrategyDataSchema-Output": {
        "properties": {
          "business_context": {
            "$ref": "#/components/schemas/BusinessContextSchema"
          },
          "audience": {
            "$ref": "#/components/schemas/AudienceSchema"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductServiceSchema"
            },
            "type": "array",
            "title": "Products"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ProductServiceSchema"
            },
            "type": "array",
            "title": "Services"
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/MarketSchema"
            },
            "type": "array",
            "title": "Markets"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/SEOGoalSchema"
            },
            "type": "array",
            "title": "Goals"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/CompetitorSchema"
            },
            "type": "array",
            "title": "Competitors"
          },
          "seo_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seo Objectives"
          },
          "content_objectives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Objectives"
          },
          "priority_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Topics"
          }
        },
        "type": "object",
        "title": "StrategyDataSchema"
      },
      "StrategyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "strategy_data": {
            "$ref": "#/components/schemas/StrategyDataSchema-Output"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "project_id",
          "current_version",
          "status",
          "strategy_data",
          "created_at",
          "updated_at"
        ],
        "title": "StrategyResponse"
      },
      "StrategyUpdateRequest": {
        "properties": {
          "strategy_data": {
            "$ref": "#/components/schemas/StrategyDataSchema-Input"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary",
            "default": "Updated strategy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "strategy_data"
        ],
        "title": "StrategyUpdateRequest"
      },
      "StrategyVersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StrategyVersionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "StrategyVersionListResponse"
      },
      "StrategyVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "strategy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Strategy Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id"
          },
          "change_summary": {
            "type": "string",
            "title": "Change Summary"
          },
          "strategy_data": {
            "$ref": "#/components/schemas/StrategyDataSchema-Output"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "strategy_id",
          "organization_id",
          "project_id",
          "version",
          "change_summary",
          "strategy_data",
          "created_at"
        ],
        "title": "StrategyVersionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}